requirements.txt
fastapi~=0.115.13
uvicorn
tensorflow~=2.19.0
pillow~=11.2.1
numpy~=2.1.3
opencv-python~=4.11.0.86
python-multipart
pydantic~=2.11.7
SQLAlchemy~=2.0.41
python-dotenv~=1.1.1